Move db traits into own file
